+++ plugins/command-code/__init__.py
@@ -1,0 +1,1 @@
+# Command Code 服务商插件